API: rename Categorical.levels to .categories
The name 'levels' was already used by a much different concept in
MultiIndex and this was too confusing, so change the name to
'categories'. Add deprecation warning if the old name is used
(in constructor or the public 'levels' attribute). The old name
'levels' is not anymore exposed unter Series.cat as it was never
part of a stable release.

See the discussion in https://github.com/pydata/pandas/issues/8074

This rename was done by search&replace in categorical.py and the
corresponding tests_categorical.py, implementing the deprecation
accessor (with a temporary 'raise Exception') and then run the the
unittests and change code until all tests pass.