Fix DataFrame.from_records error message
Make `create_block_manager_from_arrays` error message use the correct dimensions.  This is different from the situation in `create_block_manager_from_blocks` because in that method the blocks are concatenated along their first dimension, whereas in this method the arrays are concatenated along a new dimension.

Also fix `test_from_records_to_records`, which specified a misleading error message when the shape of the input data did not match the shape implied by the indices.  In particular, the error message omitted the offending dimension entirely!