BUG: Fixes #9281: fixes to tseries.tests.test_tslib.TestTimestamp
The last tests of both `test_now` and `test_today` were passing for anyone with a timezone
of UTC-5 or greater (but failed, for example, for people in US/Pacific).  The test was not
testing what the original author meant it to (which is that the times are very close
together) so I added abs(.) around the Timedeltas and also fixed the errant test.