Fix sharex/sharey behaviour together with passing in an axis
This commit fixes two things:

1.) Wrong x label visibility when using gridspec generated axis

When using gridspec, plt.gcf().get_axes() was in the wrong order
for the code handling setting the right axes label to invisible.
Also handle cases where no subplot is in the last row of a column.

2.) Don't change ax label visibility when an axis is passed in

Before, when passing in one ax object of a figure with multiple
subplots, the plot call would change the visibility of the x labels,
so that only the subplots in the last row would have xticklabels and
xlabels. This would happen even if the rest of the subplots would
not be plotted with pandas.

Now, when passing in an ax to `df.plot( ..., ax=ax)`, the `sharex`
kwarg will default to `False` and the visibility of xlabels and
xticklabels will not anymore be changed.

If you still want that, you can either do that by yourself for the
right axes in your figure or explicitly set `sharex=True`. Be aware,
that this changes the visible for all axes in the figure, not only
the one which is passed in!

If pandas creates the subplots itself (e.g. no passed in `ax` kwarg),
then the default is still `sharex=True` and the visibility changes are
applied.

Also fix some old unittests which---together with the quirks of
`_check_plot_works`, which plots the plot twice, once without an ax kwarg
and once with an ax kwarg---triggered test failures as the new
behaviour of `ax together without an explicit sharex` would not remove the
visibility of some xlabels.

Update the docstrings to explain the new sharex behaviour and also
add a warning regarding the changing of all subplots' x axis labels,
even for subplots which were not created by pandas.