DOC: hack to numpydoc to include attributes that are None (GH6100)
See GH6100. Some attributes of our classes are None, even on the
class objects themselves (so not on instances). By default, numpydoc
does not include them in the autosummary table but in a separate
plain table. But, our class.rst template does include them in the
toctree autosummary, which creates docstring pages for these. This
results in warnings "WARNING: document isn't included in any toctree".
This hack to numpydoc does include them in the autosummary in the class
docstring, removing this warning.