BUG: to_sql with datetime.time values with sqlite
I'm proposing a solution to #8341    As @jorisvandenbossche suggested,
I use a sqlite3 adapter to transform `datetime.time` objects into
string (`hh:mm:ss.ffffff`, as this what sqlalchemy... I think?)    I
added a test in the class `_TestSQLApi`, so that the solution is
tested with both sqlalchemy and the sqlite3 fallback. Thus, the result
should be consistent.

Author: Nicolas Bonnotte <nicolas.bonnotte@gmail.com>

Closes #11547 from nbonnotte/sqlite-to_sql-time and squashes the following commits:

8d449c5 [Nicolas Bonnotte] BUG: to_sql with datetime.time values with sqlite, #8341