ENH: always make ndarrays from msgpack writable
Addresses the case where 'compress' was not none. The old
implementation  would decompress the data and then call np.frombuffer
on a bytes  object. Because a bytes object is not a mutable buffer,
the resulting  ndarray had writeable=False. The new implementation
ensures that the  pandas is the only owner of this new buffer and then
sets it to mutable  without copying it. This means that we don't need
to do a copy of the  data coming in AND we can mutate it later. If we
are not the only owner  of this data then we just copy it with
np.fromstring.

Author: Joe Jevnik <joe@quantopian.com>

Closes #12359 from llllllllll/msgpack-owns-new-memory and squashes the following commits:

e896603 [Joe Jevnik] TST: adds `pandas.util.testing.patch`
84d7275 [Joe Jevnik] ENH: reimplement `_move_into_mutable_buffer` in C.
330dd76 [Joe Jevnik] ENH: updates unconvert no-copy code to be safer
830abba [Joe Jevnik] TST: adds test for memoized empty string and chars
9cd9d80 [Joe Jevnik] ENH: always make ndarrays from msgpack writable