BUG: GH12622 where pprint of Timestamp in nested structure fails
closes #12622

I suspect this is a bug in Cython (I don't think `type(A).func` should
be an `instancemethod`).  Not sure if we should use this workaround as
we encounter issues in the wild, or do this for all the functions we
can think of.

Author: Ka Wo Chen <kawoc@tepper.cmu.edu>

Closes #12629 from kawochen/BUG-FIX-12622 and squashes the following commits:

a6216ae [Ka Wo Chen] BUG: GH12622 where pprint of Timestamp in nested structure fails