DEPR: Patch to_dense behaviour for sparse.
Patches the following for `to_dense`:    1) Fix `SparseArray.to_dense`
documentation to refer to `SparseArray` and not `SparseSeries`.    2)
Deprecate the fill parameter in `SparseArray.to_dense`, as that
parameter was not being respected.    3) Deprecate the sparse_only
parameter in `SparseSeries.to_dense`, as that parameter is
inconsistent with the `to_dense` API we want, which is no parameters.
Closes #14647.

Author: gfyoung <gfyoung17@gmail.com>

Closes #14686 from gfyoung/to-dense-patch and squashes the following commits:

ad7da32 [gfyoung] BUG: Patch to_dense behaviour for sparse.