BUG: Convert uint64 in maybe_convert_numeric
Add handling for `uint64` elements in an array with the follow
behavior specifications:    1) If `uint64` and `NaN` are both
detected, the original input will be returned if `coerce_numeric`  is
`False`. Otherwise, an `Exception` is raised.    2) If `uint64` and
negative numbers are both detected, the original input be returned if
`coerce_numeric` is `False`. Otherwise, an `Exception` is raised.
Closes #14982.  Partial fix for #14983.

Author: gfyoung <gfyoung17@gmail.com>

Closes #15005 from gfyoung/maybe-convert-numeric-uint64 and squashes the following commits:

c3bd28a [gfyoung] BUG: Convert uint64 in maybe_convert_numeric