PERF: high memory in MI
closes #13904

Creates an efficient MultiIndexHashTable in cython.
This allows us to efficiently store a multi-index for fast indexing
(.get_loc() and .get_indexer()), with the current tuple-based
(and gil holding) use of the PyObject Hash Table.

This uses the pandas.tools.hashing routines to hash each of the 'values' of a MI
to a single uint64. So this makes MI more memory friendly and much
more efficient. You get these speedups, because the creation of the
hashtable is now much more efficient.

Author: Jeff Reback <jeff@reback.net>

Closes #15245 from jreback/mi and squashes the following commits:

7df6c34 [Jeff Reback] PERF: high memory in MI