BUG: The roll_quantile function now throws an exception instead of causing a segfault when quantile is out of range
closes #15463

Author: Peter Csizsek <peter.csizsek@gmail.com>

Closes #15476 from csizsek/fix-rolling-quantile-segfault and squashes the following commits:

e31e5be [Peter Csizsek] Correctly catching exception in the test for Rolling.quantile.
4eea34a [Peter Csizsek] Refactored and moved exception throwing test to a new function for Rolling.quantile().
8b1e020 [Peter Csizsek] Added a note about the Rolling.quantile bug fix to the changelog.
f39b122 [Peter Csizsek] Added a new test case to roll_quantile_test to trigger a TypeError when called with a string.
f736ca2 [Peter Csizsek] The roll_quantile function in window.pyx now raises a ValueError when the quantile value is not in [0.0, 1.0]