DOC: Patch new flake8 command grep
The grep was initially matching to "pandas," which is incorrect
because that was also matching files containing "pandas" in the name
but that were not in the main `pandas` directory (e.g. performance
test code).  This change enforces that we match to any Python files in
the main `pandas` directory.    Also picked up compatibility issue
with OSX, in which the `-r` flag does not exist.  However, `xargs`
terminates if the argument list is empty, which was the whole point of
passing in `-r` in the first place.    Follow-up to #15712

Author: gfyoung <gfyoung17@gmail.com>

Closes #15749 from gfyoung/flake8-diff-patch and squashes the following commits:

d1543b5 [gfyoung] COMPAT: Do not run xargs with -r on OSX
da57857 [gfyoung] DOC: Patch new flake8 command grep