ENH: Citing source in README file (#15856)
* Citing source in README file

For GH users who strictly or heavily use the web-view instead of a local Git, having a direct link is handy, as it does not require downloading the PDF _if_ the user wanted to go to the source of it directly. It's an alternative that allows those interested in more uploads similar to this PDF from the same author(s).

* jorisvandenbossche's feedback

I re-read the PDF and made sure the wording reflected the content presented. I also changed the source-citing so that is more friendly for .TXT files instead of Markdown or unspecified.

* Update README.txt

* English enhancement

Improved sentence structure for English speakers.